+++ CHANGELOG.md
@@ -18,3 +18,4 @@
 - version 1.0.1
 
 
+- 2026-01-31 fix: uninstall script